msg: replace parseint to parsefloat to sorting correctly

+++ routes/search/msg.js
@@ -32,7 +32,7 @@ module.exports = async function (req, res, next) {
         });
 
         p0.sort(function (a, b) {
-          return parseInt(b.MatchScore) - parseInt(a.MatchScore);
+          return parseFloat(b.MatchScore) - parseFloat(a.MatchScore);
         });
 
         res.status(200).type("json").send(p0.slice(0, parseInt((p0.length + p1.length) / 2)));
